requirements.txt
Pillow>=10.0
numpy>=1.24
pyyaml>=6.0
tqdm>=4.65
scikit-learn>=1.3
matplotlib>=3.7
torch>=2.0
transformers>=4.40
accelerate>=0.28